+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.4.1 Release (05.12.2019 - XX.12.2019) (by Tigerpanzer)
 * Added a new MOTD Manager in the bungee.yml (Now you can define the states on yourself) 
 * Now the values in bungee.yml will work
+* Arena will not be stopped anymore on starting phase when there are null players (only resetting it to waiting for players)
 
 ### 1.4.0 Release (26.10.2019 - 16.11.2019) (by Tigerpanzer)
 * Fixed incompatibility with essentials respawn